refactor: replace review functions and rename product/review functions
