+++ tests/repositories/test_product/test_repo_product_crud.py
@@ -1,9 +1,13 @@
 import pytest
 from sqlalchemy import select
+from sqlalchemy.exc import IntegrityError
 
 from app.database.models.product import Product
-from app.repositories.products import create_product, get_product_data_from_db
+from app.database.models.review import Review
+from app.repositories.products import create_product, create_product_review_repo, get_product_data_repo
 from tests.factories.products import new_product_data_factory, product_factory
+from tests.factories.reviews import new_review_data_factory
+from tests.factories.users import user_factory
 
 
 @pytest.mark.asyncio
@@ -24,8 +28,56 @@ async def test_get_product_data(db_session):
     db_session.add(new_product)
     await db_session.flush()
 
-    product_data = await get_product_data_from_db(new_product.id, db_session)
+    product_data = await get_product_data_repo(new_product.id, db_session)
 
     assert product_data.id == new_product.id
     assert product_data.name == new_product.name
     assert product_data.price == new_product.price
+
+
+@pytest.mark.asyncio
+async def test_create_review(db_session):
+    new_product = product_factory()
+    new_user = user_factory()
+    new_review_data = new_review_data_factory()
+    db_session.add_all([new_user, new_product])
+    await db_session.flush()
+
+    await create_product_review_repo(
+        new_product.id, new_review_data, new_user, db_session
+    )
+
+    review = (await db_session.execute(select(Review))).scalar_one()
+    assert review.product_id == new_product.id
+    assert review.user_id == new_user.id
+    assert review.text == new_review_data.text
+
+
+@pytest.mark.asyncio
+async def test_cannot_create_review_for_unknown_product(db_session):
+    new_product = product_factory()
+    new_user = user_factory()
+    db_session.add(new_user)
+    await db_session.flush()
+    new_review_data = new_review_data_factory()
+
+    with pytest.raises(IntegrityError):
+        await create_product_review_repo(
+            new_product.id, new_review_data, new_user, db_session
+        )
+        await db_session.flush()
+
+
+@pytest.mark.asyncio
+async def test_cannot_create_review_by_unknown_user(db_session):
+    new_product = product_factory()
+    db_session.add(new_product)
+    new_user = user_factory()
+    await db_session.flush()
+    new_review_data = new_review_data_factory()
+
+    with pytest.raises(IntegrityError):
+        await create_product_review_repo(
+            new_product.id, new_review_data, new_user, db_session
+        )
+        await db_session.flush()
